+++ index.html
@@ -86,6 +86,8 @@
     </div>
     <!-- footer section -->
     <footer>&copy; Copyright 2022 <span>AWESOME BOOKS</span></footer>
+
     <script type="module" src="./scripts/index.js"></script>
+    <script type="module" src="./modules/elements.js"></script>
   </body>
 </html>
