+++ modules/elements.js
@@ -4,3 +4,12 @@ export const bookList = document.querySelector('.book_list');
 export const addBookButton = document.querySelector('#add_book');
 export const form = document.querySelector('.book_section form');
 
+// sections
+export const showingBooksSection = document.querySelector('.showing_books');
+export const newBookSection = document.querySelector('.book_section');
+export const contactSection = document.querySelector('.contact_section');
+
+// navigation links
+export const homeLink = document.querySelector('#home_link');
+export const addBookLink = document.querySelector('#add_book_link');
+export const contactLink = document.querySelector('#contact_link');
