+++ modules/utils.js
@@ -1,0 +1,7 @@
+export function show(name, display = 'block') {
+  name.style.display = display;
+}
+
+export function hide(name) {
+  name.style.display = 'none';
+}
